# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenFuel",
    "description": "Local prototype. Synthetic seed data. No person or trip database.",
    "license": {
      "name": "AGPL-3.0-only"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/healthz": {
      "get": {
        "summary": "Health",
        "operationId": "health_healthz_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/meta": {
      "get": {
        "summary": "Meta",
        "operationId": "meta_v1_meta_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/regions": {
      "get": {
        "summary": "Regions",
        "operationId": "regions_v1_regions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/stations": {
      "get": {
        "summary": "Stations",
        "operationId": "stations_v1_stations_get",
        "parameters": [
          {
            "name": "region",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^[a-z0-9-]{1,64}$",
              "default": "demo-region",
              "title": "Region"
            }
          },
          {
            "name": "fuel",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/FuelType",
              "default": "regular"
            }
          },
          {
            "name": "payment",
            "in": "query",
            "required": false,
            "schema": {
              "$ref": "#/components/schemas/PaymentType",
              "default": "standard"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/stations/{station_id}/history": {
      "get": {
        "summary": "History",
        "operationId": "history_v1_stations__station_id__history_get",
        "parameters": [
          {
            "name": "station_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Station Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/reports": {
      "post": {
        "summary": "Report",
        "operationId": "report_v1_reports_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Report"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/audit/export": {
      "get": {
        "summary": "Audit Export",
        "operationId": "audit_export_v1_audit_export_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/v1/audit/checkpoint": {
      "get": {
        "summary": "Audit Checkpoint",
        "operationId": "audit_checkpoint_v1_audit_checkpoint_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "FuelType": {
        "type": "string",
        "enum": [
          "regular",
          "midgrade",
          "premium",
          "diesel"
        ],
        "title": "FuelType"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PaymentType": {
        "type": "string",
        "enum": [
          "standard",
          "cash",
          "credit",
          "membership"
        ],
        "title": "PaymentType"
      },
      "Report": {
        "properties": {
          "report_id": {
            "type": "string",
            "format": "uuid4",
            "title": "Report Id"
          },
          "station_id": {
            "type": "string",
            "pattern": "^[a-z0-9-]{1,64}$",
            "title": "Station Id"
          },
          "fuel_type": {
            "$ref": "#/components/schemas/FuelType"
          },
          "payment_type": {
            "$ref": "#/components/schemas/PaymentType"
          },
          "price_milli": {
            "type": "integer",
            "maximum": 15000.0,
            "minimum": 50.0,
            "title": "Price Milli"
          },
          "observed_at": {
            "type": "string",
            "format": "date-time",
            "title": "Observed At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "report_id",
          "station_id",
          "fuel_type",
          "payment_type",
          "price_milli",
          "observed_at"
        ],
        "title": "Report"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}